chore: character interface to a separate folder to reutilization
