+++ src/interfaces/Character.ts
@@ -1,0 +1,10 @@
+interface ICharacter {
+  id: number;
+  name: string;
+  thumbnail: {
+    path: string;
+    extension: string;
+  };
+}
+
+export type { ICharacter };
